Lis les notes STP

+++ Code/TheGameShip/src/model/entity/SimpleEntity.java
@@ -3,15 +3,15 @@ package model.entity;
 import java.net.URI;
 import java.util.UUID;
 
-public class SimpleEntity implements Entity{
+public class SimpleEntity implements Entity {
     private UUID id;
     private URI sprite;
-    private int X;
-    private int Y;
+    private float X;
+    private float Y;
     private String nom;
     private String type;
 
-    public SimpleEntity(URI sprite, int x, int y, String nom, String type) {
+    public SimpleEntity(URI sprite, float x, float y, String nom, String type) {
         this.id = UUID.randomUUID();
         this.sprite = sprite;
         X = x;
